move ResultDescriptor to a seperate file
