+++ operators/src/engine/operator.rs
@@ -1,4 +1,7 @@
-use super::query_processor::{TypedRasterQueryProcessor, TypedVectorQueryProcessor};
+use super::{
+    query_processor::{TypedRasterQueryProcessor, TypedVectorQueryProcessor},
+    RasterResultDescriptor, VectorResultDescriptor,
+};
 use crate::util::Result;
 use geoengine_datatypes::collections::VectorDataType;
 use geoengine_datatypes::{projection::ProjectionOption, raster::RasterDataType};
@@ -208,68 +211,6 @@ where
     }
     fn vector_sources_mut(&mut self) -> &mut [Box<dyn InitializedVectorOperator>] {
         self.vector_sources.as_mut_slice()
-    }
-}
-
-#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Serialize, Deserialize)]
-pub struct RasterResultDescriptor {
-    pub data_type: RasterDataType,
-    pub projection: ProjectionOption,
-}
-
-impl RasterResultDescriptor {
-    pub fn map<F>(self, f: F) -> Self
-    where
-        F: Fn(Self) -> Self,
-    {
-        f(self)
-    }
-
-    pub fn map_data_type<F>(mut self, f: F) -> Self
-    where
-        F: Fn(RasterDataType) -> RasterDataType,
-    {
-        self.data_type = f(self.data_type);
-        self
-    }
-
-    pub fn map_projection<F>(mut self, f: F) -> Self
-    where
-        F: Fn(ProjectionOption) -> ProjectionOption,
-    {
-        self.projection = f(self.projection);
-        self
-    }
-}
-
-#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Serialize, Deserialize)]
-pub struct VectorResultDescriptor {
-    pub data_type: VectorDataType,
-    pub projection: ProjectionOption,
-}
-
-impl VectorResultDescriptor {
-    pub fn map<F>(self, f: F) -> Self
-    where
-        F: Fn(Self) -> Self,
-    {
-        f(self)
-    }
-
-    pub fn map_data_type<F>(mut self, f: F) -> Self
-    where
-        F: Fn(VectorDataType) -> VectorDataType,
-    {
-        self.data_type = f(self.data_type);
-        self
-    }
-
-    pub fn map_projection<F>(mut self, f: F) -> Self
-    where
-        F: Fn(ProjectionOption) -> ProjectionOption,
-    {
-        self.projection = f(self.projection);
-        self
     }
 }
